% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sdm_presence.R
\name{check_sdm_presence}
\alias{check_sdm_presence}
\title{Check that the column with presences is correctly formatted}
\usage{
check_sdm_presence(.data, .col, presence_level = "presence")
}
\arguments{
\item{.data}{a \code{data.frame} or \code{tibble}, or a derived object such as an \code{sf} data.frame,
or a factor (e.g. the column with the response variable)}

\item{.col}{the column containing the presences}

\item{presence_level}{the string used to define the presence level of \code{.col}}
}
\value{
TRUE if correctly formatted
}
\description{
In \code{tidysdm}, the string defining presences should be the first level of
the response factor. This function checks that the column is correctly formatted.
}
