% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix_df.R
\name{conf_matrix_df}
\alias{conf_matrix_df}
\title{Make a confusion matrix dataframe for multiple thresholds}
\usage{
conf_matrix_df(presences, absences)
}
\arguments{
\item{presences}{Probabilities for presences}

\item{absences}{probabilities for absences}
}
\value{
A data.frame of thresholds with columns \emph{thres}, \emph{tp}, \emph{fp}, \emph{fn}, and \emph{tn}
}
\description{
Create the confusion matrix for multiple thresholds, using to optimise tss
}
\keyword{internal}
