% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control_ensemble_grid}
\alias{control_ensemble_grid}
\alias{control_ensemble_resamples}
\alias{control_ensemble_bayes}
\title{Control wrappers}
\usage{
control_ensemble_grid()

control_ensemble_resamples()

control_ensemble_bayes()
}
\value{
A \link[tune:control_grid]{tune::control_grid}, \link[tune:control_bayes]{tune::control_bayes},
or \link[tune:control_resamples]{tune::control_resamples} object.
}
\description{
Supply these light wrappers as the \code{control} argument in a
\code{\link[tune:tune_grid]{tune::tune_grid()}}, \code{\link[tune:tune_bayes]{tune::tune_bayes()}}, or \code{\link[tune:fit_resamples]{tune::fit_resamples()}}
call to return the needed elements for use in an ensemble.
These functions will return the appropriate control grid to ensure that
assessment set predictions and information on model specifications and
preprocessors, is supplied in the resampling results object!

To integrate ensemble settings with your existing control settings, note
that these functions just call the appropriate \verb{tune::control_*} function
with the arguments \verb{save_pred = TRUE, save_workflow = TRUE}.

These wrappers are equivalent to the ones used in the \code{stacks} package.
}
\seealso{
See the vignettes for examples of these functions used in context.
}
