% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg.R
\name{create_background}
\alias{create_background}
\title{Make background from points}
\usage{
create_background(
  x,
  method = c("convex_hull", "point_buffer", "ecoregion", "concave_hull", "grid_fill",
    "user_fill"),
  buffer = 1,
  max_bg = NULL,
  grid_options = NULL,
  concave_options = NULL
)
}
\arguments{
\item{concave_options}{}
}
\description{
Make background from points
}
