% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{create_prediction_grid}
\alias{create_prediction_grid}
\title{Generate a grid of values for making predictions}
\usage{
create_prediction_grid(bg, n = 10000, square = TRUE, include_polygons = FALSE)
}
\arguments{
\item{include_polygons}{}
}
\description{
Generate a grid of values for making predictions
}
