% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_formula.R
\name{form_resp}
\alias{form_resp}
\title{Get the response variable from a formula}
\usage{
form_resp(x)
}
\arguments{
\item{x}{a formula}
}
\value{
character the name of the response
}
\description{
This is the counterpart of \link[rsample:form_pred]{rsample::form_pred}.
}
\details{
Note: this might not behave well with functions such as log(y). But neither does form_pred

modified from
https://stackoverflow.com/questions/13217322/how-to-reliably-get-dependent-variable-name-from-formula-object
}
\keyword{internal}
