% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_formula.R
\name{gam_formula}
\alias{gam_formula}
\title{Create a formula for gam}
\usage{
gam_formula(object, k = 10)
}
\arguments{
\item{object}{a \link[recipes:recipe]{recipes::recipe}, already trained}

\item{k}{the \emph{k} value for the smooth}
}
\value{
a formula
}
\description{
This function takes the formula from a recipe, and turns numeric predictors
into smooths with a given k. This formula can be passed to a workflow or
workflow set when fitting a gam.
}
