% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repeat.R
\name{get_repeat}
\alias{get_repeat}
\title{Extract a simple ensemble out of a repeat ensemble}
\usage{
get_repeat(x, i)
}
\arguments{
\item{x}{a repeat ensemble object}

\item{i}{index of the simple ensemble to extract, or name of the simple
ensemble to extract}
}
\value{
a simple ensemble object
}
\description{
This function extracts a simple ensemble out of a repeat
ensemble, by index or by name.
}
\examples{
# extract the second simple ensemble out of the repeat ensemble
get_repeat(lacerta_rep_ens, i = 2)
}
