% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_vars.R
\name{get_spatial_neighbours}
\alias{get_spatial_neighbours}
\title{Get Spatial Neighbours}
\usage{
get_spatial_neighbours(
  x,
  k = 8,
  maxdist = 1e+05,
  progress = TRUE,
  parallel = 1
)
}
\arguments{
\item{x}{An \code{sf} object}

\item{k}{The number of nearest neighbours to retrieve}

\item{maxdist}{The maximum distance to get nearest neighbours from}

\item{progress}{Print progess bar?}

\item{parallel}{How many parallel cores to use.}
}
\value{
A list of neighbour indices
}
\description{
Get Spatial Neighbours
}
