% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\docType{data}
\name{horses}
\alias{horses}
\title{Coordinates of radiocarbon dates for horses}
\format{
An \code{tibble} with 1,297 rows and 3 variables:
\describe{
\item{latitude}{latitudes in degrees}
\item{longitude}{longitudes in degrees}
\item{time_bp}{time in years before present}
}
}
\usage{
horses
}
\description{
Coordinates for presences of horses from 22k to 8k YBP.
}
\keyword{datasets}
