% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km2m.R
\name{km2m}
\alias{km2m}
\title{Convert a geographic distance from km to m}
\usage{
km2m(x)
}
\arguments{
\item{x}{the number of km}
}
\value{
the number of meters
}
\description{
This function takes distance in km and converts it into meters, the
units generally used by geographic operations in \code{R}. This is a trivial
conversion, but this functions ensures that no zeroes are lost along the way!
}
\examples{
km2m(10000)
km2m(1)

}
