% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\docType{data}
\name{lacerta}
\alias{lacerta}
\title{Coordinates of presences for Iberian emerald lizard}
\format{
An \code{tibble} with 1,297 rows and 3 variables:
\describe{
\item{ID}{ids from GBIF}
\item{latitude}{latitudes in degrees}
\item{longitude}{longitudes in degrees}
}
}
\usage{
lacerta
}
\description{
Coordinates for presences of \emph{Lacerta schreiberi}. The variables are as
follows:
}
\keyword{datasets}
