% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\docType{data}
\name{lacerta_ensemble}
\alias{lacerta_ensemble}
\title{A simple ensemble for the lacerta data}
\format{
A \code{\link{simple_ensemble}} object
}
\usage{
lacerta_ensemble
}
\description{
Ensemble SDM for \emph{Lacerta schreiberi}.
}
\keyword{datasets}
