% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\docType{data}
\name{lacerta_rep_ens}
\alias{lacerta_rep_ens}
\title{A repeat ensemble for the lacerta data}
\format{
A \code{\link{repeat_ensemble}} object
}
\usage{
lacerta_rep_ens
}
\description{
Ensemble SDM for \emph{Lacerta schreiberi}, as generated in the vignette.
}
\keyword{datasets}
