% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_docs.R
\docType{data}
\name{lacertidae_background}
\alias{lacertidae_background}
\title{Coordinates of presences for lacertidae in the Iberian peninsula}
\format{
An \code{tibble} with 1,297 rows and 3 variables:
\describe{
\item{ID}{ids from GBIF}
\item{latitude}{latitudes in degrees}
\item{longitude}{longitudes in degrees}
}
}
\usage{
lacertidae_background
}
\description{
Coordinates for presences of lacertidae, used as background for the
\code{\link{lacerta}} dataset.. The variables are as follows:
}
\keyword{datasets}
