% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnet_fit.R
\name{maxnet_predict}
\alias{maxnet_predict}
\title{Wrapper to predict maxnet models}
\usage{
maxnet_predict(
  object,
  newdata,
  type = c("class", "prob"),
  maxnet_type = c("cloglog", "link", "exponential", "logistic"),
  clamp = TRUE
)
}
\description{
This function is a wrapper around the \code{predict} method for \link[maxnet:maxnet]{maxnet::maxnet},
making the function compatible with
\code{parsnip}. Users are unlikely to use this function directly.  For the
\code{parsnip} model specification for MaxEnt, see \code{\link[=maxent]{maxent()}}.
}
\details{
@param object the \link[maxnet:maxnet]{maxnet::maxnet} object
@param newdata the dataframe of new data
@param type either "prob" or "class"
@param maxnet_type the transformation used for the prediction
@param clamp logical, definign whether clamping to observed ranges should be used
@returns a tibble of predictions
}
\keyword{internal}
