% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mess_predictors.R
\name{mess_predictors}
\alias{mess_predictors}
\alias{mess_predictors,SpatRaster-method}
\alias{mess_predictors,data.frame-method}
\alias{mess_predictors,SpatRasterDataset-method}
\title{Multivariate environmental similarity surfaces (MESS)}
\usage{
mess_predictors(x, training, ...)

\S4method{mess_predictors}{SpatRaster}(x, training, .col, filename = "", ...)

\S4method{mess_predictors}{data.frame}(x, training, .col)

\S4method{mess_predictors}{SpatRasterDataset}(x, training, .col)
}
\arguments{
\item{x}{\code{\link[terra:SpatRaster-class]{terra::SpatRaster}}, \code{\link[terra:SpatRaster-class]{terra::SpatRasterDataset}} or \code{\link{data.frame}}}

\item{training}{matrix or data.frame or sf object containing the reference values; each column
should correspond to one layer of the \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object, with the exception
of the presences column defined in \code{.col} (optional).}

\item{...}{additional arguments as for \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{.col}{the column containing the presences (optional). If specified,
it is excluded when computing the MESS scores.}

\item{filename}{character. Output filename (optional)}
}
\value{
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} (data.frame) with
the MESS values.
}
\description{
Compute multivariate environmental similarity surfaces (MESS), as described
by Elith et al., 2010.
}
\details{
This function is a modified version of \code{mess} in
package \code{predicts}, with a method added to work on \code{\link[terra:SpatRaster-class]{terra::SpatRasterDataset}}.
Note that the method for \code{\link[terra:SpatRaster-class]{terra::SpatRasterDataset}} assumes that each variables
is stored as a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} with time information within \code{x}. Time
is also assumed to be in \code{years}. If these conditions are not met, it is possible
to manually extract a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} for each time step, and use
\code{mess_predictors} on those \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}s
}
\references{
Elith J., M. Kearney M., and S. Phillips, 2010. The art of
modelling range-shifting species. Methods in Ecology and Evolution
1:330-342. c("\\Sexpr[results=rd]{tools:::Rd_expr_doi(\"#1\")}",
"10.1111/j.2041-210X.2010.00036.x")\Sexpr{tools:::Rd_expr_doi("10.1111/j.2041-210X.2010.00036.x")}
}
\author{
Jean-Pierre Rossi, Robert Hijmans, Paulo van Breugel, Andrea Manica
}
