% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_thresh.R
\name{optim_thresh_sens}
\alias{optim_thresh_sens}
\title{Find threshold that gives a target sensitivity}
\usage{
optim_thresh_sens(presences, absences, sens_target)
}
\arguments{
\item{presences}{Probabilities for presences.}

\item{absences}{Provabilities for absences}

\item{sens_target}{the target sensitivity}
}
\value{
the probability threshold for the event
}
\description{
This is an internal function returns the threshold to turn probabilities into
binary classes for a given target sensitivity
}
\keyword{internal}
