% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_of_range_warning.R
\name{out_of_range_warning}
\alias{out_of_range_warning}
\title{Warn if some times are outside the range of time steps from a raster}
\usage{
out_of_range_warning(times, time_steps)
}
\arguments{
\item{times}{the times of the locations}

\item{time_steps}{the time steps from the raster}
}
\description{
This function helps making sure that, when we assign times to time_step
layers of a raster, we do not have values which are badly out of range
}
\keyword{internal}
