% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{pairs,stars-method}
\alias{pairs,stars-method}
\title{This is a wrapper around \code{graphics::pairs()} that accepts \code{stars} objects.
It is adapted from a similar function in the \code{terra} package.}
\usage{
\S4method{pairs}{stars}(
  x,
  hist = TRUE,
  cor = TRUE,
  use = "pairwise.complete.obs",
  maxcells = 1e+05,
  ...
)
}
\arguments{
\item{x}{SpatRaster}

\item{hist}{logical. If TRUE a histogram of the values is shown on the diagonal}

\item{cor}{logical. If TRUE the correlation coefficient is shown in the upper panels}

\item{use}{argument passed to the \code{\link[stats]{cor}} function}

\item{maxcells}{integer. Number of pixels to sample from each layer of a large SpatRaster}

\item{...}{additional arguments (graphical parameters)}
}
\description{
This is a wrapper around \code{graphics::pairs()} that accepts \code{stars} objects.
It is adapted from a similar function in the \code{terra} package.
}
\seealso{
\href{https://github.com/rspatial/terra/blob/a72eb63cf178c637f76859476487cd8345b529bc/R/plot.R#L115}{terra implementation}
}
