% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{po_spatial_buffer_vfold_cv}
\alias{po_spatial_buffer_vfold_cv}
\title{Spatial Cross Validation Designed for Presence-Only Data}
\usage{
po_spatial_buffer_vfold_cv(
  data,
  radius = NULL,
  buffer = 0.1,
  v = 10,
  repeats = 1,
  presence = NULL,
  pool = 0.1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Spatial Cross Validation Designed for Presence-Only Data
}
