% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_raster.R
\name{predict_raster}
\alias{predict_raster}
\alias{predict_raster.default}
\title{Make predictions for a whole raster}
\usage{
predict_raster(object, raster, ...)

\method{predict_raster}{default}(object, raster, filename = "", n = 4, ...)
}
\arguments{
\item{object}{the \code{tidymodels} object of interest}

\item{raster}{the \code{\link[terra:SpatRaster]{terra::SpatRaster}} or \code{stars} with the input data. It
has to include levels with the same names as the variables used in \code{object}}

\item{...}{parameters to be passed to the standard \code{predict()} function for
the appropriate object type (e.g. \code{metric_thresh}, \code{class_thresh} or
\code{fun}). See the documentation of the relevant \code{predict()} method for
possible parameters (e.g. \code{\link[=predict.simple_ensemble]{predict.simple_ensemble()}}).}

\item{filename}{the name of the output file raster file; this is only needed
to save large rasters, and can be left blank for rasters that can be kept
in memory.}

\item{n}{positive integer indicating how many copies the data may be in
memory at any point in time (it defaults to 4). This is used to determine
whether rasters can be processed in one go, or in chunks. If you get an out
of memory error, increase \code{n}. See \code{\link[terra:writeStart]{terra::writeStart()}} for more details.}
}
\value{
a \code{\link[terra:SpatRaster]{terra::SpatRaster}} (or \code{stars} if that is the input) with the
predictions
}
\description{
This function allows to use a raster as data to make predictions from a
variety of \code{tidymodels} objects, such as \code{\link{simple_ensemble}} or
\code{stacks::stacks}
}
\keyword{predict}
