% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_ensemble.R
\name{repeat_ensemble}
\alias{repeat_ensemble}
\title{Repeat ensemble}
\usage{
repeat_ensemble(...)
}
\arguments{
\item{...}{not used, this function just creates an empty \code{repeat_ensemble}
object. Members are added with \code{add_best_candidates()}}
}
\value{
an empty \code{repeat_ensemble}
}
\description{
An ensemble based multiple sets of pseudoabsences/background. This object
is a collection (list) of \code{\link{simple_ensemble}} objects for which predictions will
be combined in a simple way (e.g. by taking either the mean or median). Each
\code{\link{simple_ensemble}} contains the best version of a each given model type
following turning; all simple ensembles will need to have the same metric
estimated during the cv process.
}
