% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_stack.R
\name{repeat_stack}
\alias{repeat_stack}
\title{Repeat stack ensemble}
\usage{
repeat_stack(...)
}
\arguments{
\item{...}{not used, this function just creates an empty \code{repeat_stack}
object. Members are added with \code{add_stack()}}
}
\value{
an empty \code{repeat_stack}
}
\description{
An ensemble of linear stacks based multiple sets of
pseudoabsences/background. This object is a collection (list) of
\code{linear_stack} objects from package \code{stacks} for which predictions will be
combined in a simple way (e.g. by taking either the mean or median). Each
\code{linear_stack} is a weighted combination of models following turning; all
stacks will need to have the same metric estimated during the tuning process.
}
