% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrnorm.R
\name{rtrnorm_naive}
\alias{rtrnorm_naive}
\title{Sample from a truncated normal distribution}
\usage{
rtrnorm_naive(mean, sd, lower, upper)
}
\arguments{
\item{mean}{A numeric vector of means for the truncated normal distribution.}

\item{sd}{A numeric vector of standard deviations for the truncated normal
distribution.}

\item{lower}{A numeric vector of lower bounds for the truncated normal
distribution.}

\item{upper}{A numeric vector of upper bounds for the truncated normal
distribution.}
}
\value{
A numeric vector of sampled values from the truncated normal
distribution.
}
\description{
This function samples from a truncated normal distribution defined by the
mean, standard deviation, and lower and upper bounds. It uses rejection
sampling to ensure that the sampled values are within the specified bounds.
The function is vectorised, so it can handle multiple sets of parameters at
once.
}
\keyword{internal}
