% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_time_uncertainty.R
\name{sample_time_uncertainty}
\alias{sample_time_uncertainty}
\title{Sample a time when have a range or a distribution}
\usage{
sample_time_uncertainty(
  data,
  mean_sd_truncnorm_col = c("mean_time", "sd_time"),
  sd_time_units = NULL,
  min_max_unif_col = c("min_time", "max_time"),
  fixed_time_col = "fixed_time",
  lubridate_fun = c
)
}
\arguments{
\item{data}{An \code{\link[sf:sf]{sf::sf}} data frame, or a data frame with time information.
If an \code{sf} object, the geometry is ignored.}

\item{mean_sd_truncnorm_col}{a vector of length 2 with the names of the
columns with the mean and standard deviation of the truncated normal
distribution to sample from. It defaults to "mean_time" and "sd_time", but
if the column names are with different names, they can be specified here.
The "mean_time" column should be either a lubridate object, or an object
that can be converted to a lubridate object with \code{lubridate_fun}.}

\item{sd_time_units}{the units of the standard deviation of the truncated
normal'   distribution (e.g. "years" or "days"). This has to be defined
if the sd_time column exists (i.e. if some dates are being resampled from
a truncated normal).}

\item{min_max_unif_col}{a vector of length 2 with the names of the columns
with the minimum and maximum values of the uniform distribution to sample
from. It defaults to "min_time" and "max_time", but if the column names are
different, they can be specified here. The "min_time" and "max_time"
columns should be either lubridate objects, or objects that can be
converted to lubridate objects with \code{lubridate_fun}.}

\item{fixed_time_col}{The name of the column with fixed times that do not
need resampling (e.g. contemporary samples). It defaults to "fixed_time",
but if the column name is different, it can be specified here. The
"fixed_time" column should be either a lubridate object, or an object that
can be converted to a lubridate object with \code{lubridate_fun}.}

\item{lubridate_fun}{function to convert the time columns (mean, min, max and
fixed) into a lubridate object}
}
\value{
a vector of lubridate objects with the sampled times
}
\description{
The time (date) of a sample can come with uncertainty (e.g. typical of
radiocarbon dates). This function samples a time from a distribution of
possible times, which can be defined by a truncated normal distribution (with
mean and sd) or a uniform distribution (with min and max). Fixed times (i.e.
times without uncertainty) can also be included for certain samples. The
function returns a vector of sampled times.
}
\details{
Each sample (i.e. row) should only include one type of information,
i.e. either the mean and sd, or the min and max, or a fixed time.
}
