% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_data.R
\name{sdm_data}
\alias{sdm_data}
\title{Convert presence points and a background area
into an \code{sdm_data} object suitable for a
\code{tidysdm} workflow.}
\usage{
sdm_data(pres, bg, n = 500, abs = NULL, sample_options = list(), ...)
}
\arguments{
\item{pres}{Points representing species
occurrences. Should be an \code{sf} object with \code{crs}
information}

\item{bg}{A background area as an \code{sf} polygon or
a raster mask (in \code{stars} or \code{raster} format)}

\item{n}{Integer specifying the number of background or pseudo-absence
points to sample from the background as specified by \code{bg}.}

\item{abs}{Optionally an \code{sf} object with true absence data in the form of
points.}

\item{...}{}
}
\value{
A \code{sdm_data} object inheriting from a \code{tibble}
}
\description{
Convert presence points and a background area
into an \code{sdm_data} object suitable for a
\code{tidysdm} workflow.
}
