% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_metric_set.R
\name{sdm_metric_set}
\alias{sdm_metric_set}
\title{Metric set for SDM}
\usage{
sdm_metric_set(...)
}
\arguments{
\item{...}{additional metrics to be added to the
\code{\link[yardstick:metric_set]{yardstick::metric_set}}. See the help to \code{\link[yardstick:metric_set]{yardstick::metric_set()}} for
constraints on the type of metrics that can be mixed.}
}
\value{
a \code{\link[yardstick:metric_set]{yardstick::metric_set}} object.
}
\description{
This function returns a \link[yardstick:metric_set]{yardstick::metric_set} that includes \code{\link[=boyce_cont]{boyce_cont()}},
\code{\link[yardstick:roc_auc]{yardstick::roc_auc()}} and \code{\link[=tss_max]{tss_max()}}, the most commonly used metrics for
SDM.
}
\examples{
sdm_metric_set()
sdm_metric_set(accuracy)
}
