% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ensemble.R
\name{simple_ensemble}
\alias{simple_ensemble}
\title{Simple ensemble}
\usage{
simple_ensemble(...)
}
\arguments{
\item{...}{not used, this function just creates an empty \code{simple_ensemble}
object. Members are added with \code{add_best_candidates()}}
}
\value{
an empty \code{simple_ensemble}. This is a tibble with columns:
\itemize{
\item \code{wflow_id}: the name of the workflows for which the best model was
chosen
\item \code{workflow}: the trained workflow objects
\item \code{metrics}: metrics based on the crossvalidation resampling used
to tune the models
}
}
\description{
A simple ensemble is a collection of workflows for which predictions will
be combined in a simple way (e.g. by taking either the mean or median). Usually
these workflows will consists each of the best version of a given model algorithm
following tuning. The workflows are fitted to the full training dataset
before making predictions.
}
