% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_sample_pseudo_absences}
\alias{step_sample_pseudo_absences}
\title{Title}
\usage{
step_sample_pseudo_absences(
  recipe,
  bg_geometry = NULL,
  role = NA,
  trained = FALSE,
  size,
  options = list(type = "random", exact = FALSE),
  skip = FALSE,
  id = rand_id("sample_pseudo_absences")
)
}
\arguments{
\item{id}{}
}
\description{
Title
}
