% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysdm.R
\docType{package}
\name{tidysdm}
\alias{tidysdm}
\title{tidysdm}
\description{
This \code{R} library facilitates the fitting of Species Distribution
Models with \code{tidymodels}.
}
\details{
The functionalities of \code{tidysdm} are described in
Leonardi et al. (2023) \doi{10.1101/2023.07.24.550358}. Please cite it if you
use \code{tidysdm} in your research.

On its dedicated \href{https://evolecolgroup.github.io/tidysdm/}{website}, you can find
Articles giving you a step-by-step \href{https://evolecolgroup.github.io/tidysdm/articles/a0_tidysdm_overview.html}{overview of the package},
how to use \href{https://evolecolgroup.github.io/tidysdm/articles/a1_palaeodata_application.html}{\code{tidysdm} on palaeodata},
examples of \href{https://evolecolgroup.github.io/tidysdm/articles/a2_tidymodels_additions.html}{advanced modelling approaches using \code{tidymodels} features},
and a \href{https://evolecolgroup.github.io/tidysdm/articles/a3_troubleshooting.html}{troubleshooting guide for when models fail}.
There is also a
\href{https://evolecolgroup.github.io/tidysdm/dev/}{development version} of the site
updated for the \code{dev} version (on the top left, the version number is in
red, and will be in the format x.x.x.9xxx, indicating it is a
development version).
}
