% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y2d.R
\name{y2d}
\alias{y2d}
\title{Convert a time interval from years to days}
\usage{
y2d(x)
}
\arguments{
\item{x}{the number of years of the interval}
}
\value{
a \code{difftime} object (in days)
}
\description{
This function takes takes a time interval in years and converts into days,
the unit commonly used in time operations in \code{R}. The simple conversion
x * 365 does not work for large number of years, due to the presence of
leap years.
}
\examples{
y2d(1)
y2d(1000)

}
