% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_repeat.R
\name{add_repeat}
\alias{add_repeat}
\alias{add_repeat.default}
\alias{add_repeat.simple_ensemble}
\alias{add_repeat.list}
\title{Add repeat(s) to a repeated ensemble}
\usage{
add_repeat(x, rep, ...)

\method{add_repeat}{default}(x, rep, ...)

\method{add_repeat}{simple_ensemble}(x, rep, ...)

\method{add_repeat}{list}(x, rep, ...)
}
\arguments{
\item{x}{a \link{repeat_ensemble} to which repeat(s) will be added}

\item{rep}{a repeat, as a single \code{\link{simple_ensemble}}, or a list of
\code{\link{simple_ensemble}} objects}

\item{...}{not used at the moment.}
}
\value{
a \link{repeat_ensemble} with additional repeat(s)
}
\description{
This function adds repeat(s) to a \code{\link{repeat_ensemble}} object, where each
repeat is a \code{\link{simple_ensemble}}. All repeats must contain the same members,
selected using the same metric.
}
