% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_repeated_ensemble.R
\name{autoplot.repeat_ensemble}
\alias{autoplot.repeat_ensemble}
\title{Plot the results of a repeat ensemble}
\usage{
\method{autoplot}{repeat_ensemble}(object, ...)
}
\arguments{
\item{object}{A \code{\link{repeat_ensemble}} whose elements have results.}

\item{...}{Other options to pass to \code{autoplot()}. Currently unused.}
}
\value{
A ggplot object.
}
\description{
This \code{autoplot()} method plots performance metrics for each repeat,
ranked using one of the metrics.
}
\examples{
autoplot(lacerta_rep_ens)

}
\keyword{ensemble}
