% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_collinear.R
\name{filter_collinear}
\alias{filter_collinear}
\alias{filter_collinear.default}
\alias{filter_collinear.SpatRaster}
\alias{filter_collinear.data.frame}
\title{Filter to retain only variables that have low collinearity}
\usage{
filter_collinear(
  x,
  cutoff = NULL,
  verbose = FALSE,
  names = TRUE,
  to_keep = NULL,
  method = "caret",
  maxcell = Inf
)

\method{filter_collinear}{default}(
  x,
  cutoff = NULL,
  verbose = FALSE,
  names = TRUE,
  to_keep = NULL,
  method = "caret",
  max_cells = Inf
)

\method{filter_collinear}{SpatRaster}(
  x,
  cutoff = 0.7,
  verbose = FALSE,
  names = TRUE,
  to_keep = NULL,
  method = "caret",
  max_cells = Inf
)

\method{filter_collinear}{data.frame}(
  x,
  cutoff = NULL,
  verbose = FALSE,
  names = TRUE,
  to_keep = NULL,
  method = "caret",
  max_cells = Inf
)
}
\arguments{
\item{x}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object, a data.frame (with only numeric
variables), or a correlation matrix}

\item{cutoff}{A numeric value used as a threshold to remove variables. For, "caret",
it is the pair-wise absolute correlation cutoff, which defaults to 0.7. For "vifstep"
and "vifcor", it is the variable inflation factor, which defaults to 10}

\item{verbose}{A boolean whether additional information should be provided
on the screen}

\item{names}{a logical; should the column names be returned \code{TRUE} or
the column index \code{FALSE})?}

\item{to_keep}{A vector of variable names that we want to force in the set
(note that the function will return an error if the correlation among any of
those variables is higher than the cutoff).}

\item{method}{character. One of "caret", "vifcor" or "vifstep"}

\item{maxcell}{positive integer. The maximum number of cells to be used. If this is smaller than ncell(x), a regular sample of x is used}
}
\value{
A vector of names of columns that are below the correlation threshold
(when \code{names = TRUE}), otherwise a vector of indices. Note that the indices
are only for numeric variables (i.e. if factors are present, the indices do
not take them into account).
}
\description{
This method finds a subset of variables that have low collinearity. It provides
three methods: \code{cor_caret}, a stepwise approach to remove variables with a pairwise correlation
above a given cutoff (based on the algorithm in \code{caret::findCorrelation});
\code{vifstep}, a stepwise appraoch
such that all have a correlation
below a certain cutoff. There are methods for \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}, and
\code{\link{data.frame}}. Only numeric variables will be considered. For \code{data.frame}, only numeric variables will be
considered.
The algorithm is based on \code{caret::findCorrelation}, using the \code{exact} option.
The absolute values of pair-wise correlations are considered. If two
variables have a high correlation, the function looks at the mean absolute
correlation of each variable and removes the variable with the largest mean
absolute correlation.
}
\details{
There are several function in the package \code{subselect}
that can also be used to accomplish
the same goal but tend to retain more predictors.
}
