% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_high_cor.R
\name{filter_high_cor}
\alias{filter_high_cor}
\title{Filter to retain only variables below a given correlation threshold}
\usage{
filter_high_cor(x, cutoff = 0.7, verbose = FALSE, names = TRUE)
}
\arguments{
\item{x}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object}

\item{cutoff}{A numeric value for the pair-wise absolute correlation cutoff}

\item{verbose}{A boolean for printing the details}

\item{names}{a logical; should the column names be returned \code{TRUE} or
the column index \code{FALSE})?}
}
\value{
A vector of names of columns that are below the correlation threhold
(when \code{names = TRUE}), otherwise a vector of indeces.
}
\description{
This function estimates the correlation matrix among variables in a
\link[terra:SpatRaster-class]{terra::SpatRaster} and then returns a vector of variable names or indeces
that are below a certain cutoff.
The algorithm is based on \code{caret::findCorrelation}, using the \code{exact} option.
The absolute values of pair-wise correlations are considered. If two
variables have a high correlation, the function looks at the mean absolute
correlation of each variable and removes the variable with the largest mean
absolute correlation.
}
\details{
There are several function in the \code{subselect}
that can also be used to accomplish
the same goal but tend to retain more predictors.
}
