% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_high_cor.R
\name{filter_high_cor}
\alias{filter_high_cor}
\alias{filter_high_cor.default}
\alias{filter_high_cor.SpatRaster}
\alias{filter_high_cor.data.frame}
\alias{filter_high_cor.matrix}
\title{Deprecated: Filter to retain only variables below a given correlation
threshold}
\usage{
filter_high_cor(x, cutoff = 0.7, verbose = FALSE, names = TRUE, to_keep = NULL)

\method{filter_high_cor}{default}(x, cutoff = 0.7, verbose = FALSE, names = TRUE, to_keep = NULL)

\method{filter_high_cor}{SpatRaster}(x, cutoff = 0.7, verbose = FALSE, names = TRUE, to_keep = NULL)

\method{filter_high_cor}{data.frame}(x, cutoff = 0.7, verbose = FALSE, names = TRUE, to_keep = NULL)

\method{filter_high_cor}{matrix}(x, cutoff = 0.7, verbose = FALSE, names = TRUE, to_keep = NULL)
}
\arguments{
\item{x}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object, a data.frame (with only numeric
variables), or a correlation matrix}

\item{cutoff}{A numeric value for the pair-wise absolute correlation cutoff}

\item{verbose}{A boolean for printing the details}

\item{names}{a logical; should the column names be returned \code{TRUE} or the
column index \code{FALSE})?}

\item{to_keep}{A vector of variable names that we want to force in the set
(note that the function will return an error if the correlation among any
of those variables is higher than the cutoff).}
}
\value{
A vector of names of columns that are below the correlation
threshold (when \code{names = TRUE}), otherwise a vector of indices. Note
that the indices are only for numeric variables (i.e. if factors are
present, the indices do not take them into account).
}
\description{
THIS FUNCTION IS DEPRECATED. USE \code{filter_collinear} with \code{method=cor_caret}
instead
}
\details{
This method finds a subset of variable such that all have a correlation below
a certain cutoff. There are methods for \code{\link[terra:SpatRaster-class]{terra::SpatRaster}},
\code{\link{data.frame}}, and to work directly on a correlation matrix that was
previously estimated. For \code{data.frame}, only numeric variables will be
considered. The algorithm is based on \code{caret::findCorrelation}, using the
\code{exact} option. The absolute values of pair-wise correlations are considered.
If two variables have a high correlation, the function looks at the mean
absolute correlation of each variable and removes the variable with the
largest mean absolute correlation.

There are several function in the package \code{subselect} that can also be used
to accomplish the same goal but tend to retain more predictors.
}
