% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mess_predictors.R
\name{mess_predictors}
\alias{mess_predictors}
\title{Multivariate environmental similarity surfaces (MESS)}
\usage{
mess_predictors(x, training, .col, ...)
}
\arguments{
\item{x}{\code{\link[terra:SpatRaster-class]{terra::SpatRaster}}, \code{\link[terra:SpatRaster-class]{terra::SpatRasterDataset}} or \code{\link{data.frame}}}

\item{.col}{the column containing the presences (optional). If specified,
it is excluded from the clamping.}

\item{...}{additional arguments as for \code{\link[terra:writeRaster]{terra::writeRaster()}}}

\item{v}{matrix or data.frame containing the reference values; each column
should correspond to one layer of the \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object.}

\item{filename}{character. Output filename (optional)}
}
\value{
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} (data.frame) with
the MESS values.
}
\description{
Compute multivariate environmental similarity surfaces (MESS), as described
by Elith et al., 2010. This function is a modified version of \code{mess} in
package \code{predicts}, with a method added to work on \code{\link[terra:SpatRaster-class]{terra::SpatRasterDataset}}.
}
\details{
\code{v} can be obtained for a set of points using
\code{\link[terra:extract]{terra::extract()}} .
}
\references{
Elith J., M. Kearney M., and S. Phillips, 2010. The art of
modelling range-shifting species. Methods in Ecology and Evolution
1:330-342. c("\\Sexpr[results=rd]{tools:::Rd_expr_doi(\"#1\")}",
"10.1111/j.2041-210X.2010.00036.x")\Sexpr{tools:::Rd_expr_doi("10.1111/j.2041-210X.2010.00036.x")}
}
\author{
Jean-Pierre Rossi, Robert Hijmans, Paulo van Breugel, Andrea Manica
}
