% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_raster.R
\name{predict_raster}
\alias{predict_raster}
\alias{predict_raster.default}
\title{Make predictions for a whole raster}
\usage{
predict_raster(object, raster, ...)

\method{predict_raster}{default}(object, raster, ...)
}
\arguments{
\item{object}{the \code{\link{tidymodels}} object of interest}

\item{raster}{the \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} with the input data. It has to include
levels with the same names as the variables used in \code{object}}

\item{...}{parameters to be passed to the standard \code{predict()} function
for the appropriate object type.}

\item{metric_thresh}{a vector of length 2 giving a metric and its threshold,
which will be used to prune
which models in the ensemble will be used for the prediction. The metric's threshold needs to match
the value used in \code{\link{calib_class_thresh}}. Examples are c("accuracy",0.8) or c("boyce_cont",0.7).}
}
\value{
a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} with the predictions
}
\description{
This function allows to use a raster as data to make predictions from a
variety of \link{tidymodels} objects, such as \code{\link{simple_ensemble}} or \code{stacks::linear_stack}
}
