% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_simple_ensemble.R
\name{prob_to_binary}
\alias{prob_to_binary}
\title{simple function to convert probability to binary classes}
\usage{
prob_to_binary(x, thresh, class_levels)
}
\arguments{
\item{x}{a vector of probabilities}

\item{thresh}{the threshold to convert to binary}

\item{class_levels}{the binary levels}
}
\value{
a vector of binary values
}
\description{
simple function to convert probability to binary classes
}
\keyword{internal}
