% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_data.R
\name{sdm_data}
\alias{sdm_data}
\title{Convert presence points and a background area
into an \code{sdm_data} object suitable for a
\code{tidysdm} workflow.}
\usage{
sdm_data(presence_pnts, bg, n = 500, ...)
}
\arguments{
\item{presence_pnts}{Points representing species
occurrences. Should be an \code{sf} object with \code{crs}
information}

\item{bg}{A background area as an \code{sf} polygon or
a raster mask (in \code{stars} or \code{raster} format)}

\item{n}{Number of grid cells to divide the background
area into. These grid cells will be used for spatial
resampling (e.g. with \code{spatialsample} package)}

\item{...}{}
}
\value{
A \code{sdm_data} object inheriting from a \code{tibble}
}
\description{
Convert presence points and a background area
into an \code{sdm_data} object suitable for a
\code{tidysdm} workflow.
}
