% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysdm-package.R
\docType{package}
\name{tidysdm-package}
\alias{tidysdm}
\alias{tidysdm-package}
\title{tidysdm: Species Distribution Models with Tidymodels}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fit species distribution models (SDMs) using the 'tidymodels' framework, which provides a standardised interface to define models and process their outputs. 'tidysdm' expands 'tidymodels' by providing methods for spatial objects, models and metrics specific to SDMs, as well as a number of specialised functions to process occurrences for contemporary and palaeo datasets. The full functionalities of the package are described in Leonardi et al. (2024) \doi{10.1111/2041-210X.14406}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EvolEcolGroup/tidysdm}
  \item \url{https://evolecolgroup.github.io/tidysdm/}
  \item Report bugs at \url{https://github.com/EvolEcolGroup/tidysdm/issues}
}

}
\author{
\strong{Maintainer}: Andrea Manica \email{am315@cam.ac.uk}

Authors:
\itemize{
  \item Michela Leonardi
  \item Margherita Colucci
  \item Andrea Vittorio Pozzi
  \item Eleanor M.L. Scerri
}

Other contributors:
\itemize{
  \item Ben Tupper [contributor]
  \item Ondrej Pelanek [contributor]
}

}
\keyword{internal}
